# Optimize resource sorting performance