Trades slight complexity for better performance.

Please note that since `BenchmarkInitPushContext` does not set `CreationTimestamp` for the resources, the time of the resources are the same, which causes the `name/namespace` comparison to be performed when sorting, which is the main reason why the benchmark shows a significant improvement in performance. In a production environment, there are generally not too many resources with the same `CreationTimestamp`, so the overall performance improvement of this optimization is limited.

```
#  BenchmarkInitPushContext
goos: darwin
goarch: amd64
pkg: istio.io/istio/pilot/pkg/xds
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                                              │   old.txt    │               new.txt               │
                                              │    sec/op    │    sec/op     vs base               │
InitPushContext/gateways-12                      12.30m ± 5%   10.97m ±  2%  -10.76% (p=0.002 n=6)
InitPushContext/gateways-shared-12              10.808m ± 5%   9.828m ± 12%        ~ (p=0.093 n=6)
InitPushContext/knative-gateway-12               1.308m ± 1%   1.346m ±  6%        ~ (p=0.394 n=6)
InitPushContext/http-12                          224.1µ ± 6%   156.1µ ±  8%  -30.34% (p=0.002 n=6)
InitPushContext/tcp-12                           226.2µ ± 5%   156.4µ ±  5%  -30.84% (p=0.002 n=6)
InitPushContext/tls-12                           221.6µ ± 4%   154.2µ ±  5%  -30.40% (p=0.002 n=6)
InitPushContext/auto-12                          223.1µ ± 4%   153.9µ ±  8%  -31.03% (p=0.002 n=6)
InitPushContext/strict-12                        11.14µ ± 3%   11.03µ ±  6%        ~ (p=0.589 n=6)
InitPushContext/disabled-12                      11.22µ ± 1%   11.07µ ±  4%        ~ (p=0.240 n=6)
InitPushContext/externalname-12                  133.9µ ± 3%   129.6µ ±  4%   -3.26% (p=0.026 n=6)
InitPushContext/telemetry-api-12                 279.0µ ± 1%   206.1µ ±  3%  -26.11% (p=0.002 n=6)
InitPushContext/virtualservice-12                496.0µ ± 2%   428.6µ ±  2%  -13.60% (p=0.002 n=6)
InitPushContext/authorizationpolicy-12           54.08µ ± 6%   54.63µ ±  3%        ~ (p=0.589 n=6)
InitPushContext/serviceentry-workloadentry-12    433.6µ ± 2%   365.8µ ±  2%  -15.65% (p=0.002 n=6)
geomean                                          285.3µ        241.9µ        -15.21%

                                              │   old.txt    │               new.txt               │
                                              │     B/op     │     B/op      vs base               │
InitPushContext/gateways-12                     7.222Mi ± 0%   6.471Mi ± 0%  -10.40% (p=0.002 n=6)
InitPushContext/gateways-shared-12              6.268Mi ± 0%   5.517Mi ± 0%  -11.98% (p=0.002 n=6)
InitPushContext/knative-gateway-12              999.0Ki ± 0%   998.8Ki ± 0%   -0.03% (p=0.002 n=6)
InitPushContext/http-12                         168.4Ki ± 1%   118.3Ki ± 0%  -29.74% (p=0.002 n=6)
InitPushContext/tcp-12                          168.5Ki ± 1%   118.4Ki ± 0%  -29.73% (p=0.002 n=6)
InitPushContext/tls-12                          168.6Ki ± 1%   118.5Ki ± 0%  -29.72% (p=0.002 n=6)
InitPushContext/auto-12                         168.6Ki ± 1%   118.5Ki ± 0%  -29.72% (p=0.002 n=6)
InitPushContext/strict-12                       5.817Ki ± 0%   5.817Ki ± 0%        ~ (p=0.810 n=6)
InitPushContext/disabled-12                     5.816Ki ± 0%   5.817Ki ± 0%        ~ (p=0.870 n=6)
InitPushContext/externalname-12                 71.27Ki ± 1%   71.31Ki ± 1%        ~ (p=0.818 n=6)
InitPushContext/telemetry-api-12                171.2Ki ± 1%   121.9Ki ± 0%  -28.77% (p=0.002 n=6)
InitPushContext/virtualservice-12               483.5Ki ± 0%   433.0Ki ± 0%  -10.44% (p=0.002 n=6)
InitPushContext/authorizationpolicy-12          73.50Ki ± 0%   73.50Ki ± 0%        ~ (p=0.381 n=6)
InitPushContext/serviceentry-workloadentry-12   266.6Ki ± 1%   216.1Ki ± 0%  -18.92% (p=0.002 n=6)
geomean                                         198.7Ki        168.5Ki       -15.20%

                                              │   old.txt   │               new.txt                │
                                              │  allocs/op  │  allocs/op   vs base                 │
InitPushContext/gateways-12                     80.16k ± 0%   55.46k ± 0%  -30.81% (p=0.002 n=6)
InitPushContext/gateways-shared-12              71.81k ± 0%   47.22k ± 0%  -34.25% (p=0.002 n=6)
InitPushContext/knative-gateway-12              11.24k ± 0%   11.24k ± 0%   -0.04% (p=0.002 n=6)
InitPushContext/http-12                         2.790k ± 2%   1.187k ± 0%  -57.46% (p=0.002 n=6)
InitPushContext/tcp-12                          2.800k ± 1%   1.187k ± 0%  -57.60% (p=0.002 n=6)
InitPushContext/tls-12                          2.800k ± 1%   1.187k ± 0%  -57.61% (p=0.002 n=6)
InitPushContext/auto-12                         2.801k ± 1%   1.187k ± 0%  -57.63% (p=0.002 n=6)
InitPushContext/strict-12                        102.0 ± 0%    102.0 ± 0%        ~ (p=1.000 n=6) ¹
InitPushContext/disabled-12                      102.0 ± 0%    102.0 ± 0%        ~ (p=1.000 n=6) ¹
InitPushContext/externalname-12                  729.0 ± 0%    729.0 ± 0%        ~ (p=1.000 n=6) ¹
InitPushContext/telemetry-api-12                3.090k ± 1%   1.498k ± 0%  -51.51% (p=0.002 n=6)
InitPushContext/virtualservice-12               4.711k ± 1%   3.106k ± 0%  -34.07% (p=0.002 n=6)
InitPushContext/authorizationpolicy-12           107.0 ± 0%    107.0 ± 0%        ~ (p=1.000 n=6) ¹
InitPushContext/serviceentry-workloadentry-12   4.806k ± 1%   3.194k ± 0%  -33.54% (p=0.002 n=6)
geomean                                         2.412k        1.597k       -33.76%
¹ all samples are equal
```