@istio-testing 2024-06-25
@wulianglongrd: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
gencheck_istio | 5735e893f9b82d95f7226a13bca46c4060d20a54 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/51684/gencheck_istio/1805458157229576192) | true | `/test gencheck`
unit-tests-arm64_istio | 5735e893f9b82d95f7226a13bca46c4060d20a54 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/51684/unit-tests-arm64_istio/1805458169791516672) | true | `/test unit-tests-arm64`
lint_istio | 5735e893f9b82d95f7226a13bca46c4060d20a54 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/51684/lint_istio/1805458166444462080) | true | `/test lint`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes-sigs/prow](https://github.com/kubernetes-sigs/prow/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->