@wulianglongrd 2024-06-25
/test integ-pilot-istiodremote